(PE-12299) Only rebuild Solaris 11 repo for new packages
Only rebuild the Solaris 11 repo when it doesn't contain the needed
package.

+++ manifests/osfamily/solaris.pp
@@ -43,48 +43,55 @@ class puppet_agent::osfamily::solaris(
       contain puppet_agent::prepare::package
 
       $pkgrepo_dir = '/etc/puppetlabs/installer/solaris.repo'
-
-      # Make sure the pkg publishers are all available.  Broken
-      # publisher entries will stop the installation process.
-      # This must happen before removing any packages.
-      # We rely on the puppetlabs.com publisher previously being
-      # setup (during the initial install).
-      exec { 'puppet_agent ensure pkg publishers are available':
-        command   => 'pkg refresh',
-        path      => '/bin:/usr/bin:/sbin:/usr/sbin',
-        logoutput => 'on_failure',
-        notify    => Exec['puppet_agent remove existing repo'],
+      $publisher = 'puppetlabs.com'
+      $arch = $::architecture ? {
+        /^sun4[uv]$/ => 'sparc',
+        default      => 'i386',
       }
+      $pkg_name = basename($package_file_name, ".${arch}.p5p")
 
       exec { 'puppet_agent remove existing repo':
         command   => "rm -rf '${pkgrepo_dir}'",
         path      => '/bin:/usr/bin:/sbin:/usr/sbin',
         onlyif    => "test -d ${pkgrepo_dir}",
         logoutput => 'on_failure',
-        notify    => Exec['puppet_agent create repo'],
-      }
-
+        unless    => "pkgrepo list -p ${publisher} -s ${pkgrepo_dir} ${pkg_name}",
+      } ~>
       exec { 'puppet_agent create repo':
         command     => "pkgrepo create ${pkgrepo_dir}",
         path        => '/bin:/usr/bin:/sbin:/usr/sbin',
         unless      => "test -f ${pkgrepo_dir}/pkg5.repository",
         logoutput   => 'on_failure',
-        notify      => Exec['puppet_agent set publisher'],
         refreshonly => true,
-      }
-
+      } ~>
       exec { 'puppet_agent set publisher':
-        command     => "pkgrepo set -s ${pkgrepo_dir} publisher/prefix=puppetlabs.com",
+        command     => "pkgrepo set -s ${pkgrepo_dir} publisher/prefix=${publisher}",
         path        => '/bin:/usr/bin:/sbin:/usr/sbin',
         logoutput   => 'on_failure',
         refreshonly => true,
-      }
-
+      } ~>
+      exec { 'puppet_agent ensure pkgrepo is up-to-date':
+        command     => "pkgrepo refresh -s ${pkgrepo_dir}",
+        path        => '/bin:/usr/bin:/sbin:/usr/sbin',
+        logoutput   => 'on_failure',
+        refreshonly => true,
+      } ~>
       exec { 'puppet_agent copy packages':
-        command   => "pkgrecv -s file:///opt/puppetlabs/packages/${package_file_name} -d ${pkgrepo_dir} '*'",
-        path      => '/bin:/usr/bin:/sbin:/usr/sbin',
-        require   => Exec['puppet_agent set publisher'],
-        logoutput => 'on_failure',
+        command     => "pkgrecv -s file:///opt/puppetlabs/packages/${package_file_name} -d ${pkgrepo_dir} '*'",
+        path        => '/bin:/usr/bin:/sbin:/usr/sbin',
+        logoutput   => 'on_failure',
+        refreshonly => true,
+      } ~>
+      # Make sure the pkg publishers are all available.  Broken
+      # publisher entries will stop the installation process.
+      # This must happen before removing any packages.
+      # We rely on the puppetlabs.com publisher previously being
+      # setup (during the initial install).
+      exec { 'puppet_agent ensure pkg publishers are available':
+        command     => "pkg refresh ${publisher}",
+        path        => '/bin:/usr/bin:/sbin:/usr/sbin',
+        logoutput   => 'on_failure',
+        refreshonly => true,
       }
 
       if versioncmp("${::clientversion}", '4.0.0') < 0 {
